+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 #Changelog
 
+##Version 0.3.0
+* Added support for built-in django migrations.
+* Fixed an error with south migrations.
+* It is now possible to set max_length to something other than length. (max_length will default to length if not set.)
+
 ##Version 0.2.1
 * Fixed an error in docstrings.